+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 - Add a filter button to the statistics view
 - Smooth the per-activity statistics graphs to make it easier to look at them
+- Round values to 0 digits in the track graph view for the HeartRate and Elevation charts